requirements.txt
# =====================================================
# CORE BACKEND
# =====================================================
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-multipart==0.0.9
requests==2.32.3

# =====================================================
# DATABASE
# =====================================================
sqlalchemy==2.0.31
psycopg2-binary==2.9.9

# =====================================================
# OCR & PDF PROCESSING
# =====================================================
pdfplumber==0.11.4
pypdf==4.2.0
opencv-python-headless==4.10.0.84
pillow==10.4.0

# =====================================================
# NLP / EMBEDDINGS / RAG
# =====================================================
fastembed
faiss-cpu==1.8.0
numpy==1.26.4

# =====================================================
# DASHBOARD
# =====================================================
streamlit==1.36.0

# =====================================================
# ENV / CONFIG
# =====================================================
python-dotenv==1.0.1

# =====================================================
# TESTING
# =====================================================
pytest==8.2.2
